minor tweaks

+++ icinga/start.sh
@@ -14,7 +14,7 @@ log_dir=${log_dir-'/var/log/icinga'}
 state_dir=${state_dir-"${base_dir}/run"}
 cache_dir=${cache_dir-'/var/cache/icinga'}
 lib_dir=${lib_dir-'/var/lib/icinga'}
-checkresult_dir=${checkresult_dir-'/var/lib/icinga/spool/checkresults'}
+checkresult_dir=${checkresult_dir-"${lib_dir}/spool/checkresults"}
 extra_args=${extra_args-''}
 
 for dir_to_clean in "${cache_dir}" "${checkresult_dir}"; do
@@ -35,12 +35,18 @@ done
 mkdir -p /run/php
 
 # Export static content to nginx container, if data volume mount is present
-test -d /icinga_data && rsync --delete -va /usr/share/icinga/htdocs/ /icinga_data/htdocs
-test -L /icinga_data/htdocs/jquery && rm /icinga_data/htdocs/jquery
-test -L /icinga_data/htdocs/jquery-ui && rm /icinga_data/htdocs/jquery-ui
-test -d /icinga_data && rsync --delete -va /usr/share/javascript/jquery/ /icinga_data/htdocs/jquery
-test -d /icinga_data && rsync --delete -va /usr/share/javascript/jquery-ui/ /icinga_data/htdocs/jquery-ui
-#test -d /icinga_data && rsync --delete -va /etc/icinga/ /icinga_data/conf   # hope there are no secrets in there
-test -d /icinga_data && (cd /icinga_data/ && tar zxvf /stylesheets.tar.gz)   # aaaaaaaaaaaaaaaargh
+if [ -d /icinga_data ]; then
+    rsync --delete -va /usr/share/icinga/htdocs/ /icinga_data/htdocs
+    test -L /icinga_data/htdocs/jquery && rm /icinga_data/htdocs/jquery
+    test -L /icinga_data/htdocs/jquery-ui && rm /icinga_data/htdocs/jquery-ui
+    rsync --delete -va /usr/share/javascript/jquery/ /icinga_data/htdocs/jquery
+    rsync --delete -va /usr/share/javascript/jquery-ui/ /icinga_data/htdocs/jquery-ui
+    #rsync --delete -va /etc/icinga/ /icinga_data/conf   # hope there are no secrets in there
+    test -f /stylesheets.tar.gz && (cd /icinga_data/ && tar zxvf /stylesheets.tar.gz)   # aaaaaaaaaaaaaaaargh - extract files preserved in setup.sh
+else
+    echo ""
+    echo "$0: WARNING: /icinga_data (data volume) not present - can't export web files"
+    echo ""
+fi
 
 exec /usr/bin/supervisord -c /etc/supervisor/conf.d/supervisord.conf
